+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ## v0.3
 
 ### v0.3.4
+- 🍏 enhance `need` default err message
 - 🍏 add `bytes.concat([bytesA, bytesB, bytesC])`
 
 ### v0.3.3